pyproject.toml
[tool.poetry]
name = "emersit-parse"
version = "0.1.0"
description = ""
authors = ["Kamil Ayupov <ayupovkamil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
aiogram = "^3.3.0"
python-dotenv = "^1.0.1"
fake-useragent = "^1.4.0"
aiosqlite = "^0.19.0"
sqlalchemy = "^2.0.25"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
